requirements.txt
django==4.1
djangorestframework
psycopg2-binary
whitenoise
google-api-python-client
pytube
python-dotenv